+++ package.json
@@ -10,6 +10,7 @@
     "deploy": "gh-pages -d dist"
   },
   "dependencies": {
+    "bootstrap": "^5.2.2",
     "gh-pages": "^4.0.0",
     "vue": "^3.2.37"
   },
